requirements.txt
black
click
mypy
pydantic
pytest
python-pdfbox
